requirements.txt
# === Django et backend web ===
Django==4.2.22
djangorestframework
django_celery_results
celery
redis
flower

# === IA & NLP / Embeddings ===
sentence-transformers>=2.6.1,<3.0.0
torch>=2.1.0,<2.2.0
transformers>=4.31.0,<5.0.0
keybert
stanza
bert-score

# === PDF & OCR ===
PyMuPDF
pdfminer.six
paddleocr
paddlepaddle==3.0.0
argostranslate

# === Data science de base ===
numpy>=1.25.0,<2.0.0
pandas>=1.5,<2.0
scikit-learn>=1.3,<1.4

# === Recherche vectorielle (FAISS) ===
faiss-cpu==1.11.0

# === Utilitaires ===
langdetect
requests
uvicorn[standard]
